fix return value objects

+++ Domain/ValueObjects/Balance.cs
@@ -2,10 +2,10 @@
 
 public record struct Balance
 {
-    public float Value { get; set; }
-    public Balance(float value)
+    public double Value { get; set; }
+    public Balance(double value)
     {
-        if (value >= 0.0f)
+        if (value >= 0.0)
             Value = value;
         else
             throw new Exception("Balance invalid!");
